# internal/appointments/ports/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package http

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"scheduler/internal/appointments/domain"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// AppointmentDetailResponse defines model for AppointmentDetailResponse.
type AppointmentDetailResponse struct {
	// AppointmentUuid UUID of the appointment
	AppointmentUuid  AppointmentUUID `json:"appointmentUuid"`
	DealerShipId     string          `json:"dealerShipId"`
	EstimatedEndTime time.Time       `json:"estimatedEndTime"`
	ServiceBayId     string          `json:"serviceBayId"`
	ServiceType      string          `json:"serviceType"`
	StartTime        time.Time       `json:"startTime"`

	// Status Status of the appointment
	Status       AppointmentStatus `json:"status"`
	TechnicianId string            `json:"technicianId"`

	// VehicleUuid UUID of the vehicle
	VehicleUuid VehicleUUID `json:"vehicleUuid"`
}

// AppointmentStatus Status of the appointment
type AppointmentStatus = domain.AppointmentStatus

// AppointmentUUID UUID of the appointment
type AppointmentUUID = domain.AppointmentUUID

// ConfirmAppointmentResponse defines model for ConfirmAppointmentResponse.
type ConfirmAppointmentResponse struct {
	// AppointmentUuid UUID of the appointment
	AppointmentUuid AppointmentUUID `json:"appointmentUuid"`
}

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  *string `json:"error_slug,omitempty"`
	Message    *string `json:"message,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Details *[]ErrorDetail `json:"details,omitempty"`
	Message string         `json:"message"`
	Slug    string         `json:"slug"`
}

// GetReservationResponse defines model for GetReservationResponse.
type GetReservationResponse struct {
	Appointment AppointmentDetailResponse `json:"appointment"`
	ExpiredAt   time.Time                 `json:"expiredAt"`

	// ReservationUuid UUID of the reservation
	ReservationUuid ReservationUUID `json:"reservationUuid"`
}

// ReservationUUID UUID of the reservation
type ReservationUUID = domain.ReservationUUID

// VehicleUUID UUID of the vehicle
type VehicleUUID = domain.VehicleUUID

// CancelAppointmentParams defines parameters for CancelAppointment.
type CancelAppointmentParams struct {
	XUserId string `json:"X-User-Id"`
}

// GetReservationParams defines parameters for GetReservation.
type GetReservationParams struct {
	XUserId string `json:"X-User-Id"`
}

// ConfirmAppointmentParams defines parameters for ConfirmAppointment.
type ConfirmAppointmentParams struct {
	XUserId string `json:"X-User-Id"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// CancelAppointment Cancel a confirmed appointment
	// (POST /appointments/{appointmentUuid}/cancel)
	CancelAppointment(ctx echo.Context, appointmentUuid AppointmentUUID, params CancelAppointmentParams) error
	// GetReservation Get a reservation together with its appointment
	// (GET /reservations/{reservationUuid})
	GetReservation(ctx echo.Context, reservationUuid ReservationUUID, params GetReservationParams) error
	// ConfirmAppointment Confirm a pending appointment using its reservation
	// (POST /reservations/{reservationUuid}/confirm)
	ConfirmAppointment(ctx echo.Context, reservationUuid ReservationUUID, params ConfirmAppointmentParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CancelAppointment converts echo context to params.
func (w *ServerInterfaceWrapper) CancelAppointment(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "appointmentUuid" -------------
	var appointmentUuid AppointmentUUID

	err = runtime.BindStyledParameterWithOptions("simple", "appointmentUuid", ctx.Param("appointmentUuid"), &appointmentUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter appointmentUuid: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params CancelAppointmentParams

	headers := ctx.Request().Header
	// ------------- Required header parameter "X-User-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-User-Id")]; found {
		var XUserId string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-User-Id, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-User-Id", valueList[0], &XUserId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-User-Id: %s", err))
		}

		params.XUserId = XUserId
	} else {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Header parameter X-User-Id is required, but not found"))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CancelAppointment(ctx, appointmentUuid, params)
	return err
}

// GetReservation converts echo context to params.
func (w *ServerInterfaceWrapper) GetReservation(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "reservationUuid" -------------
	var reservationUuid ReservationUUID

	err = runtime.BindStyledParameterWithOptions("simple", "reservationUuid", ctx.Param("reservationUuid"), &reservationUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter reservationUuid: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetReservationParams

	headers := ctx.Request().Header
	// ------------- Required header parameter "X-User-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-User-Id")]; found {
		var XUserId string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-User-Id, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-User-Id", valueList[0], &XUserId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-User-Id: %s", err))
		}

		params.XUserId = XUserId
	} else {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Header parameter X-User-Id is required, but not found"))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetReservation(ctx, reservationUuid, params)
	return err
}

// ConfirmAppointment converts echo context to params.
func (w *ServerInterfaceWrapper) ConfirmAppointment(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "reservationUuid" -------------
	var reservationUuid ReservationUUID

	err = runtime.BindStyledParameterWithOptions("simple", "reservationUuid", ctx.Param("reservationUuid"), &reservationUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter reservationUuid: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ConfirmAppointmentParams

	headers := ctx.Request().Header
	// ------------- Required header parameter "X-User-Id" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-User-Id")]; found {
		var XUserId string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-User-Id, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-User-Id", valueList[0], &XUserId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-User-Id: %s", err))
		}

		params.XUserId = XUserId
	} else {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Header parameter X-User-Id is required, but not found"))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ConfirmAppointment(ctx, reservationUuid, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/reservations/:reservationUuid/confirm", wrapper.ConfirmAppointment, options.OperationMiddlewares["confirmAppointment"]...)
	router.GET(options.BaseURL+"/reservations/:reservationUuid", wrapper.GetReservation, options.OperationMiddlewares["getReservation"]...)
	router.POST(options.BaseURL+"/appointments/:appointmentUuid/cancel", wrapper.CancelAppointment, options.OperationMiddlewares["cancelAppointment"]...)

}

type CancelAppointmentRequestObject struct {
	AppointmentUuid AppointmentUUID `json:"appointmentUuid"`
	Params          CancelAppointmentParams
}

type CancelAppointmentResponseObject interface {
	VisitCancelAppointmentResponse(w http.ResponseWriter) error
}

type CancelAppointment204Response struct {
}

func (response CancelAppointment204Response) VisitCancelAppointmentResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type CancelAppointment400JSONResponse ErrorResponse

func (response CancelAppointment400JSONResponse) VisitCancelAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type CancelAppointment401JSONResponse ErrorResponse

func (response CancelAppointment401JSONResponse) VisitCancelAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type CancelAppointment404JSONResponse ErrorResponse

func (response CancelAppointment404JSONResponse) VisitCancelAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type CancelAppointment409JSONResponse ErrorResponse

func (response CancelAppointment409JSONResponse) VisitCancelAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type CancelAppointmentdefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response CancelAppointmentdefaultJSONResponse) VisitCancelAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetReservationRequestObject struct {
	ReservationUuid ReservationUUID `json:"reservationUuid"`
	Params          GetReservationParams
}

type GetReservationResponseObject interface {
	VisitGetReservationResponse(w http.ResponseWriter) error
}

type GetReservation200JSONResponse GetReservationResponse

func (response GetReservation200JSONResponse) VisitGetReservationResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetReservation401JSONResponse ErrorResponse

func (response GetReservation401JSONResponse) VisitGetReservationResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetReservation404JSONResponse ErrorResponse

func (response GetReservation404JSONResponse) VisitGetReservationResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetReservationdefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response GetReservationdefaultJSONResponse) VisitGetReservationResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointmentRequestObject struct {
	ReservationUuid ReservationUUID `json:"reservationUuid"`
	Params          ConfirmAppointmentParams
}

type ConfirmAppointmentResponseObject interface {
	VisitConfirmAppointmentResponse(w http.ResponseWriter) error
}

type ConfirmAppointment200JSONResponse ConfirmAppointmentResponse

func (response ConfirmAppointment200JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointment400JSONResponse ErrorResponse

func (response ConfirmAppointment400JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointment401JSONResponse ErrorResponse

func (response ConfirmAppointment401JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointment404JSONResponse ErrorResponse

func (response ConfirmAppointment404JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointment409JSONResponse ErrorResponse

func (response ConfirmAppointment409JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointment410JSONResponse ErrorResponse

func (response ConfirmAppointment410JSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(410)
	_, err := buf.WriteTo(w)
	return err
}

type ConfirmAppointmentdefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response ConfirmAppointmentdefaultJSONResponse) VisitConfirmAppointmentResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// CancelAppointment Cancel a confirmed appointment
	// (POST /appointments/{appointmentUuid}/cancel)
	CancelAppointment(ctx context.Context, request CancelAppointmentRequestObject) (CancelAppointmentResponseObject, error)
	// GetReservation Get a reservation together with its appointment
	// (GET /reservations/{reservationUuid})
	GetReservation(ctx context.Context, request GetReservationRequestObject) (GetReservationResponseObject, error)
	// ConfirmAppointment Confirm a pending appointment using its reservation
	// (POST /reservations/{reservationUuid}/confirm)
	ConfirmAppointment(ctx context.Context, request ConfirmAppointmentRequestObject) (ConfirmAppointmentResponseObject, error)
}

type StrictHandlerFunc func(ctx echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// CancelAppointment operation middleware
func (sh *strictHandler) CancelAppointment(ctx echo.Context, appointmentUuid AppointmentUUID, params CancelAppointmentParams) error {
	var request CancelAppointmentRequestObject

	request.AppointmentUuid = appointmentUuid
	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CancelAppointment(ctx.Request().Context(), request.(CancelAppointmentRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CancelAppointment")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CancelAppointmentResponseObject); ok {
		return validResponse.VisitCancelAppointmentResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetReservation operation middleware
func (sh *strictHandler) GetReservation(ctx echo.Context, reservationUuid ReservationUUID, params GetReservationParams) error {
	var request GetReservationRequestObject

	request.ReservationUuid = reservationUuid
	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetReservation(ctx.Request().Context(), request.(GetReservationRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReservation")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetReservationResponseObject); ok {
		return validResponse.VisitGetReservationResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// ConfirmAppointment operation middleware
func (sh *strictHandler) ConfirmAppointment(ctx echo.Context, reservationUuid ReservationUUID, params ConfirmAppointmentParams) error {
	var request ConfirmAppointmentRequestObject

	request.ReservationUuid = reservationUuid
	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ConfirmAppointment(ctx.Request().Context(), request.(ConfirmAppointmentRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ConfirmAppointment")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ConfirmAppointmentResponseObject); ok {
		return validResponse.VisitConfirmAppointmentResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}